refactor: move contexts to only wrap game
